size_t周りの修正

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/07 15:45:28 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/15 19:50:42 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/18 19:02:02 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,13 +15,13 @@
 void	*ft_calloc(size_t n, size_t size)
 {
 	char	*ptr;
-	int		i;
+	size_t	i;
 
 	ptr = (char *)malloc(size * n);
 	if (!ptr)
 		return (NULL);
 	i = 0;
-	while (i < (int)(size * n))
+	while (i < (size * n))
 	{
 		ptr[i] = '\0';
 		i++;
