+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 20:20:32 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/14 19:31:33 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/18 19:00:47 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t size)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*ptr1;
 	unsigned char	*ptr2;
 
 	i = 0;
 	ptr1 = (unsigned char *)s1;
 	ptr2 = (unsigned char *)s2;
-	while (i < (int)size)
+	while (i < size)
 	{
 		if (ptr1[i] != ptr2[i])
 			return (ptr1[i] - ptr2[i]);
