+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 14:47:57 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/07 18:23:49 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/18 18:50:16 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 void	*ft_memset(void *str, int c, size_t size)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*ptr;
 	unsigned char	value;
 
 	i = 0;
 	ptr = (unsigned char *)str;
 	value = (unsigned char)c;
-	while (i < (int)size)
+	while (i < size)
 	{
 		ptr[i] = value;
 		i++;
