+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 20:41:54 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/15 19:00:17 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/18 19:09:23 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_strnstr(const char *s1, const char *s2, size_t size)
 
 	p1 = s1;
 	p2 = s2;
-	while (*p1 && *p2 && (int)size >= 1)
+	while (*p1 && *p2 && size >= 1)
 	{
 		if (*p1 == *p2)
 		{
@@ -38,3 +38,27 @@ char	*ft_strnstr(const char *s1, const char *s2, size_t size)
 	else
 		return ((char *)(p1 - (p2 - s2)));
 }
+
+//test10
+//#include <stdio.h>
+
+//int main(void)
+//{
+//	char haystack[30] = "aaabcabcd";
+
+//	printf("ft_strnstr = %s\n", ft_strnstr(haystack, "abcd", 9));
+//	printf("%s\n", haystack + 5);
+//}
+
+//test2
+//#include <stdio.h>
+
+//int main(void)
+//{
+//	char haystack[30] = "aaabcabcd";
+//	char needle[10] = "aabc";
+//	//char * empty = (char*)"";
+
+//	printf("ft_strnstr = %s\n", ft_strnstr(haystack, needle, -1));
+//	printf("%s\n", haystack + 1);
+//}
